Refactoring of soft and force servers deletion
We shouldn't use force deletion for servers by default as it is not
default behavior for nova. Force deletion can be used when test
requires physical deletion of servers or in cleanup as we should
physically delete servers before the next tests in order to minimize
problems with the lack of memory or quotas.

Change-Id: I2f53bef02e5af4a52d9f81567ff5ce6e5d1edea1