Increase the USERDATA_EXECUTING_TIMEOUT for LM tests
Now it's equal to 5*60 seconds that's not enough even for
complete boot of Ubuntu on a server launched against
cluster with virtual nodes. The UBUNTU_BOOT_COMPLETE_TIMEOUT
used by another tests was was set to 10 * 60 seconds (approved
by specific script running). So we can add an extra time to it for
relevant USERDATA_EXECUTING_TIMEOUT.

Change-Id: Ic2192a53d367ac4226277c74b4442da6a62873a2