Mark nova tests on workload as destructive
If cleanup of nova workload tests fails, environment can hang up and as
result tests after that will fail. So tests on nova workload were
marked as destructive.
But as we should also check that nova works under workload, we should
try to delete servers in tests (not in cleanup) because cleanup is not
performed for destructive tests.

Change-Id: Ia916022e5c6a1f0c83d9fdbff85dc4fa87dd4286